refactor(Prompts): move form components to forms/ subdirectory
