+++ client/src/components/Prompts/forms/PromptLabelsForm.tsx
@@ -3,8 +3,10 @@ import { Input } from '@librechat/client';
 import { Cross1Icon } from '@radix-ui/react-icons';
 import type { TPrompt } from 'librechat-data-provider';
 import { useUpdatePromptLabels } from '~/data-provider';
+import { useLocalize } from '~/hooks';
 
 const PromptForm = ({ selectedPrompt }: { selectedPrompt?: TPrompt }) => {
+  const localize = useLocalize();
   const [labelInput, setLabelInput] = useState<string>('');
   const [labels, setLabels] = useState<string[]>([]);
   const updatePromptLabelsMutation = useUpdatePromptLabels();
@@ -36,7 +38,9 @@ const PromptForm = ({ selectedPrompt }: { selectedPrompt?: TPrompt }) => {
         onChange={handleInputChange}
         onKeyPress={handleKeyPress}
       />
-      <h3 className="rounded-t-lg border border-gray-300 px-4 text-base font-semibold">Labels</h3>
+      <h3 className="rounded-t-lg border border-gray-300 px-4 text-base font-semibold">
+        {localize('com_ui_labels')}
+      </h3>
       <div className="mb-4 flex w-full flex-row flex-wrap rounded-b-lg border border-gray-300 p-4">
         {labels.length ? (
           labels.map((label, index) => (
@@ -59,7 +63,7 @@ const PromptForm = ({ selectedPrompt }: { selectedPrompt?: TPrompt }) => {
             </label>
           ))
         ) : (
-          <label className="rounded-full border px-2">No Labels</label>
+          <label className="rounded-full border px-2">{localize('com_ui_no_labels')}</label>
         )}
       </div>
     </>
