+++ client/src/components/Prompts/forms/index.ts
@@ -1,0 +1,4 @@
+export { default as PromptForm } from './PromptForm';
+export { default as CreatePromptForm } from './CreatePromptForm';
+export { default as VariableForm } from './VariableForm';
+export { default as PromptLabelsForm } from './PromptLabelsForm';
